Move agent runtime interface to agents package
