+++ hephaestus/automation/address_review.py
@@ -25,8 +25,14 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
 
+from hephaestus.agents.runtime import (
+    add_agent_argument,
+    is_codex,
+    resume_codex_session,
+    run_codex_session,
+)
+
 from ._review_utils import find_pr_for_issue
-from .agent_runtime import add_agent_argument, is_codex, resume_codex_session, run_codex_session
 from .claude_models import implementer_model
 from .claude_timeouts import address_review_claude_timeout
 from .curses_ui import CursesUI, ThreadLogManager
