+++ hephaestus/automation/implementer.py
@@ -23,19 +23,19 @@ from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, cast
 
-from hephaestus.github.rate_limit import (
-    detect_claude_usage_cap,
-    detect_rate_limit,
-    wait_until,
-)
-
-from .agent_runtime import (
+from hephaestus.agents.runtime import (
     add_agent_argument,
     is_codex,
     resume_codex_session,
     run_codex_session,
     run_codex_text,
 )
+from hephaestus.github.rate_limit import (
+    detect_claude_usage_cap,
+    detect_rate_limit,
+    wait_until,
+)
+
 from .claude_invoke import parse_review_verdict
 from .claude_models import implementer_model, reviewer_model
 from .claude_timeouts import implementer_claude_timeout
