+++ hephaestus/automation/planner.py
@@ -21,9 +21,9 @@ from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 from pathlib import Path
 from typing import Any
 
+from hephaestus.agents.runtime import add_agent_argument, is_codex, run_codex_text
 from hephaestus.github.rate_limit import wait_until
 
-from .agent_runtime import add_agent_argument, is_codex, run_codex_text
 from .claude_invoke import parse_review_verdict, scan_quota_reset
 from .claude_models import advise_model, learn_model, planner_model, reviewer_model
 from .claude_timeouts import planner_claude_timeout
